requirements.txt
langchain
streamlit
openai
tiktoken
python-dotenv
unstructured
pinecone-client
pypdf
sentence_transformers
accelerate